[Refactor] UpcomingDateDetailContentView로 리팩토링 (#30)
